requirements/dev.txt
jupyter < 2
mike < 2
mkdocs < 2
mkdocs-click < 0.7
mkdocs-jupyter < 0.25
mkdocstrings-python < 2
nbmake < 2
pre-commit < 4
pytest < 8
pytest-cov < 5
pytest-mock < 4
pytest-timeout < 3
pytest-xdist < 4

